+++ packages/common/web/src/index.js
@@ -90,7 +90,9 @@ export {
   getAnchorBlockData,
   mergeBlockData,
   isAtomicBlockFocused,
-  removeBlock,
+  updateEntityData,
+  replaceWithEmptyBlock,
+  deleteBlock,
 } from './Utils/draftUtils';
 export { default as Version } from './Utils/versioningUtils';
 export { isiOS } from './Utils/isiOS';