+++ packages/editor/web/src/RichContentEditor/handleBackspaceCommand.js
@@ -1,12 +1,12 @@
 import { RichUtils, EditorState } from 'draft-js';
-import { isAtomicBlockFocused, removeBlock } from 'wix-rich-content-common';
+import { isAtomicBlockFocused, replaceWithEmptyBlock } from 'wix-rich-content-common';
 import removeBlockAdjacentToAtomic from './atomicBlockRemovalUtil';
 
 export default editorState => {
   const selection = editorState.getSelection();
 
   if (isAtomicBlockFocused(editorState)) {
-    return removeBlock(editorState, selection.getAnchorKey());
+    return replaceWithEmptyBlock(editorState, selection.getAnchorKey());
   }
 
   if (!selection.isCollapsed() || selection.getAnchorOffset() || selection.getFocusOffset()) {